pyproject.toml
[project]
name = "neshast-backend"
version = "0.1.0"
description = "FOSS Event"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.1.7",
    "djangorestframework-simplejwt>=5.5.0",
    "djangorestframework>=3.15.2",
    "python-decouple>=3.8",
    "ruff>=0.11.2",
    "psycopg2-binary>=2.9.10",
    "django-stubs>=5.1.3",
    "celery[redis]>=5.4.0",
    "django-storages[s3]>=1.14.5",
    "django-cors-headers>=4.7.0",
    "pillow>=11.1.0",
    "django-extensions>=4.1",
    "factory-boy>=3.3.3",
    "identiconify>=0.1.0",
    "httpx>=0.28.1",
    "pydantic[email]>=2.11.5",
    "drf-nested-routers>=0.94.2",
    "drf-spectacular>=0.28.0",
]
